Improved RNG issues
For each phase of searching for objects, one extra object is selected for viability if allowed based on number of objects present. For example if the player needs one generator, two will be available. However the player can never need more than are present. Since that would be a hard lock for progression.